Adjusted chat feature
